backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-filter==23.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.5.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.2
